Move CSS flags into their own folder
