+++ packages/alfa-css/src/tsconfig.json
@@ -31,7 +31,6 @@
     "./syntax/parser.ts",
     "./syntax/token.ts",
     "./value/box.ts",
-    "./value/contain.ts",
     "./unit/converter.ts",
     "./unit/index.ts",
     "./unit/unit.ts",
@@ -47,6 +46,8 @@
     "./value/color/named.ts",
     "./value/color/rgb.ts",
     "./value/color/system.ts",
+    "./value/flags/contain.ts",
+    "./value/flags/index.ts",
     "./value/image/gradient/gradient.ts",
     "./value/image/gradient/index.ts",
     "./value/image/gradient/item/hint.ts",
