+++ packages/alfa-css/src/value/flags/index.ts
@@ -1,0 +1,1 @@
+export * from "./contain.js";
