+++ packages/alfa-css/src/value/index.ts
@@ -1,7 +1,7 @@
 export * from "./box.js";
-export * from "./contain.js";
 export * from "./collection/index.js";
 export * from "./color/index.js";
+export * from "./flags/index.js";
 export * from "./image/index.js";
 export * from "./numeric/index.js";
 export * from "./position/index.js";
